feat(STOCK-20) removed Swagger and refactor IndicatorService logic
