+++ src/main/java/lissa/trading/analytics/service/dto/IndicatorsDto.java
@@ -1,11 +1,10 @@
 package lissa.trading.analytics.service.dto;
 
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Schema(description = "Тело ответа содержащее технические индикаторы")
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
