--- src/main/java/lissa/trading/analytics/service/exception/CalculationException.java
@@ -1,7 +1,0 @@
-package lissa.trading.analytics.service.exception;
-
-public class CalculationException extends RuntimeException{
-    public CalculationException(String message){
-        super(message);
-    }
-}
