+++ src/main/java/lissa/trading/analytics/service/model/CandleInterval.java
@@ -1,5 +1,9 @@
 package lissa.trading.analytics.service.model;
 
 public enum CandleInterval {
-    CANDLE_INTERVAL_5_MIN, CANDLE_INTERVAL_HOUR, CANDLE_INTERVAL_DAY, CANDLE_INTERVAL_WEEK, CANDLE_INTERVAL_MONTH
+    CANDLE_INTERVAL_5_MIN,
+    CANDLE_INTERVAL_HOUR,
+    CANDLE_INTERVAL_DAY,
+    CANDLE_INTERVAL_WEEK,
+    CANDLE_INTERVAL_MONTH
 }
