+++ src/main/resources/application-local.yml
@@ -5,7 +5,8 @@ security:
   internal:
     token: Iq6Vu8Nz7Xa3Rl9Cm5YbTc2WpDf4Oj5ZfGn1Kh3
   tinkoff:
-    token: t.FIi_ZiShdpwlsg3HFmt4we1IlpNjkoDntBuPtRJXjZQ9OQWdFZTas40CCIzsiBe1bHLoH-L1lV1axVHkV3WTTg
+    # Paste your tinkoff-invest token for test this service
+    token:
 
 integration:
   rest:
